+++ CHANGES.md
@@ -13,6 +13,8 @@
 
 - [CHANGE] ビルド周りを完全にリニューアルする
   - @melpon
+- [CHANGE] raspberry-pi-os_armv6 と raspberry-pi-os_armv7 を削除
+  - @melpon
 - [UPDATE] CLI11 を 2.4.2 に上げる
   - @voluntas @torikizi
 - [UPDATE] SDL を 2.30.3 に上げる